requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
idna==3.6
psycopg2==2.9.9
requests==2.31.0
soupsieve==2.5
Unidecode==1.3.8
urllib3==2.2.0
